+++ ecs_fargate/tests/test_unit.py
@@ -7,8 +7,8 @@ import os
 
 import mock
 import pytest
-import simplejson as json
 
+from datadog_checks.dev.http import MockResponse
 from datadog_checks.ecs_fargate import FargateCheck
 
 HERE = os.path.dirname(os.path.abspath(__file__))
@@ -63,25 +63,13 @@ def check():
     return FargateCheck('ecs_fargate', {}, {})
 
 
-class MockResponse:
-    def __init__(self, json_data, status_code):
-        self.json_data = json_data
-        self.status_code = status_code
-
-    def json(self):
-        return self.json_data
-
-
 def mocked_requests_get(*args, **kwargs):
     if args[0].endswith("/metadata"):
-        fpath = os.path.join(HERE, 'fixtures', 'metadata.json')
+        return MockResponse(file_path=os.path.join(HERE, 'fixtures', 'metadata.json'))
     elif args[0].endswith("/stats"):
-        fpath = os.path.join(HERE, 'fixtures', 'stats.json')
+        return MockResponse(file_path=os.path.join(HERE, 'fixtures', 'stats.json'))
     else:
-        return MockResponse(None, 404)
-
-    with open(fpath) as f:
-        return MockResponse(json.loads(f.read()), 200)
+        return MockResponse(status_code=404)
 
 
 def mocked_get_tags(entity, _):
@@ -143,8 +131,11 @@ def test_failing_check(check, instance, aggregator):
     """
     Testing fargate metadata endpoint error.
     """
-    with mock.patch('datadog_checks.ecs_fargate.ecs_fargate.requests.get', return_value=MockResponse("{}", 500)):
-        check.check(instance)
+    check.instance = instance
+    with mock.patch(
+        'datadog_checks.ecs_fargate.ecs_fargate.requests.get', return_value=MockResponse('{}', status_code=500)
+    ):
+        check.check({})
 
     aggregator.assert_service_check("fargate_check", status=FargateCheck.CRITICAL, tags=INSTANCE_TAGS, count=1)
 
@@ -153,8 +144,11 @@ def test_invalid_response_check(check, instance, aggregator):
     """
     Testing invalid fargate metadata payload.
     """
-    with mock.patch('datadog_checks.ecs_fargate.ecs_fargate.requests.get', return_value=MockResponse("{}", 200)):
-        check.check(instance)
+    check.instance = instance
+    with mock.patch(
+        'datadog_checks.ecs_fargate.ecs_fargate.requests.get', return_value=MockResponse('{}', status_code=200)
+    ):
+        check.check({})
 
     aggregator.assert_service_check("fargate_check", status=FargateCheck.WARNING, tags=INSTANCE_TAGS, count=1)
 
@@ -163,10 +157,11 @@ def test_successful_check(check, instance, aggregator):
     """
     Testing successful fargate check.
     """
+    check.instance = instance
     with mock.patch('datadog_checks.ecs_fargate.ecs_fargate.requests.get', side_effect=mocked_requests_get):
         with mock.patch("datadog_checks.ecs_fargate.ecs_fargate.get_tags", side_effect=mocked_get_tags):
             with mock.patch("datadog_checks.ecs_fargate.ecs_fargate.c_is_excluded", side_effect=mocked_is_excluded):
-                check.check(instance)
+                check.check({})
 
     aggregator.assert_service_check("fargate_check", status=FargateCheck.OK, tags=INSTANCE_TAGS, count=1)
 
@@ -240,10 +235,16 @@ def test_config(test_case, extra_config, expected_http_kwargs):
     with mock.patch('datadog_checks.base.utils.http.requests') as r:
         r.get.return_value = mock.MagicMock(status_code=200)
 
-        check.check(instance)
+        check.check({})
 
         http_wargs = dict(
-            auth=mock.ANY, cert=mock.ANY, headers=mock.ANY, proxies=mock.ANY, timeout=mock.ANY, verify=mock.ANY
+            auth=mock.ANY,
+            cert=mock.ANY,
+            headers=mock.ANY,
+            proxies=mock.ANY,
+            timeout=mock.ANY,
+            verify=mock.ANY,
+            allow_redirects=mock.ANY,
         )
         http_wargs.update(expected_http_kwargs)
         r.get.assert_called_with('http://169.254.170.2/v2/metadata', **http_wargs)